generate-data/requirements.txt
aiohttp
aiosignal
anyio
async-timeout
attrs
bleach
certifi
charset-normalizer
click
click-log
colorama
dataclasses-json
deprecation
docutils
dotty-dict
frozenlist
gitdb
GitPython
gotrue
greenlet
h11
httpcore
httpx
idna
importlib-metadata
invoke
jaraco.classes
keyring
langchain
marshmallow
marshmallow-enum
more-itertools
multidict
mypy-extensions
numexpr
numpy
openai
openapi-schema-pydantic
packaging
pkginfo
postgrest
pydantic
Pygments
python-dateutil
python-dotenv
python-gitlab
python-semantic-release
PyYAML
readme-renderer
realtime
regex
requests
requests-toolbelt
rfc3986
semver
six
smmap
sniffio
SQLAlchemy
storage3
StrEnum
supabase
supafunc
tenacity
tiktoken
tomlkit
tqdm
twine
typing-inspect
typing_extensions
urllib3<1.37
webencodings
websockets
yarl
zipp
